Cleanup of unnecessary classes/code (kitty proyect's files).
