+++ app/src/main/AndroidManifest.xml
@@ -16,12 +16,12 @@
         android:theme="@style/Theme.Mobile2_Scaffolding"
         tools:targetApi="31">
         <activity
-            android:name=".ui.quiz.QuizActivity"
+            android:name=".ui.QuizActivity"
             android:exported="false"
             android:label="@string/title_activity_quiz"
             android:theme="@style/Theme.Mobile2_Scaffolding" />
         <activity
-            android:name=".ui.heroDuel.HeroDuelActivity"
+            android:name=".ui.HeroDuelActivity"
             android:exported="false"
             android:label="@string/title_activity_hero_fight"
             android:theme="@style/Theme.Mobile2_Scaffolding" />
@@ -31,12 +31,12 @@
             android:label="@string/title_activity_new_game"
             android:theme="@style/Theme.Mobile2_Scaffolding" />
         <activity
-            android:name=".ui.collection.CollectionActivity"
+            android:name=".ui.CollectionActivity"
             android:exported="false"
             android:label="@string/title_activity_collection"
             android:theme="@style/Theme.Mobile2_Scaffolding" />
         <activity
-            android:name=".ui.heroDetail.HeroDetailActivity"
+            android:name=".ui.HeroDetailActivity"
             android:exported="false"
             android:label="@string/title_activity_hero_detail"
             android:theme="@style/Theme.Mobile2_Scaffolding" />
