--- app/src/main/java/ar/edu/unlam/mobile2/HttpClientModule.kt
@@ -1,53 +1,0 @@
-package ar.edu.unlam.mobile2
-
-import android.content.Context
-import coil.ImageLoader
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ViewModelComponent
-import okhttp3.OkHttpClient
-import retrofit2.Retrofit
-import retrofit2.Retrofit.Builder
-import retrofit2.converter.gson.GsonConverterFactory
-import java.security.SecureRandom
-import java.security.cert.X509Certificate
-import javax.net.ssl.HostnameVerifier
-import javax.net.ssl.SSLContext
-import javax.net.ssl.TrustManager
-import javax.net.ssl.X509TrustManager
-
-@Module
-@InstallIn(ViewModelComponent::class)
-class HttpClientModule {
-
-    @Provides
-    fun retrofitBuilder(): Builder {
-        return Retrofit.Builder()
-            .client(OkHttpClient.Builder().apply { ignoreAllSSLErrors() }.build())
-            .addConverterFactory(GsonConverterFactory.create())
-    }
-
-    fun imageLoader(context: Context): ImageLoader {
-        return ImageLoader.Builder(context)
-            .okHttpClient(OkHttpClient.Builder().apply { ignoreAllSSLErrors() }.build())
-            .build()
-    }
-
-    fun OkHttpClient.Builder.ignoreAllSSLErrors(): OkHttpClient.Builder {
-        val naiveTrustManager = object : X509TrustManager {
-            override fun getAcceptedIssuers(): Array<X509Certificate> = arrayOf()
-            override fun checkClientTrusted(certs: Array<X509Certificate>, authType: String) = Unit
-            override fun checkServerTrusted(certs: Array<X509Certificate>, authType: String) = Unit
-        }
-
-        val insecureSocketFactory = SSLContext.getInstance("TLSv1.2").apply {
-            val trustAllCerts = arrayOf<TrustManager>(naiveTrustManager)
-            init(null, trustAllCerts, SecureRandom())
-        }.socketFactory
-
-        sslSocketFactory(insecureSocketFactory, naiveTrustManager)
-        hostnameVerifier(HostnameVerifier { _, _ -> true })
-        return this
-    }
-}
