--- app/src/main/java/ar/edu/unlam/mobile2/data/KittiesRepo.kt
@@ -1,6 +1,0 @@
-package ar.edu.unlam.mobile2.data
-
-interface KittiesRepo {
-
-    suspend fun getNewKitty(): String
-}
