--- app/src/main/java/ar/edu/unlam/mobile2/data/KittyApiModel.kt
@@ -1,11 +1,0 @@
-package ar.edu.unlam.mobile2.data
-
-import ar.edu.unlam.mobile2.domain.hero.Image
-
-data class KittyApiModel(
-    val id: String,
-    val url: String,
-    val width: Int,
-    val height: Int,
-
-)
