+++ app/src/main/java/ar/edu/unlam/mobile2/ui/utilities/HeroState.kt
@@ -1,4 +1,4 @@
-package ar.edu.unlam.mobile2.ui.heroDetail
+package ar.edu.unlam.mobile2.ui.utilities
 
 import ar.edu.unlam.mobile2.domain.hero.DataHero
