+++ app/src/main/java/ar/edu/unlam/mobile2/ui/viewmodel/HeroDetailViewModelImp.kt
@@ -1,4 +1,4 @@
-package ar.edu.unlam.mobile2.ui.collection
+package ar.edu.unlam.mobile2.ui.viewmodel
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -6,25 +6,33 @@ import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import ar.edu.unlam.mobile2.data.repository.IHeroRepository
+import ar.edu.unlam.mobile2.domain.hero.DataHero
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class CollectionViewModelImp @Inject constructor(private val repo: IHeroRepository):ViewModel() {
-    var heroList by mutableStateOf(HeroListState())
+class HeroDetailViewModelImp @Inject constructor(private val repo: IHeroRepository): ViewModel() {
+    /*
+    var hero by mutableStateOf(HeroState())
+        private set
+    */
+    var hero by mutableStateOf(DataHero())
         private set
 
-    init {
+    fun getHero(id: Int) {
         viewModelScope.launch {
-            heroList = heroList.copy(
+            /*
+            hero = hero.copy(
                 isLoading = true
             )
-            heroList = heroList.copy(
-                heroList = repo.getAllHero(),
+            hero = hero.copy(
+                dataHero = repo.getHero(id),
                 isLoading = false
             )
+
+             */
+            hero = repo.getHero(id)
         }
     }
-
 }
